refactor: change reference route helper to custom hook
