+++ apps/web/src/services/asset-inventory/components/CloudServiceDetail.vue
@@ -41,8 +41,8 @@ import {
 } from '@/lib/excel-export';
 import { downloadExcelByExportFetcher } from '@/lib/helper/file-download-helper';
 import { MENU_ID } from '@/lib/menu/config';
-import { referenceFieldFormatter } from '@/lib/reference/referenceFieldFormatter';
 import type { Reference } from '@/lib/reference/type';
+import { useReferenceFieldFormatter } from '@/lib/reference/use-reference-field-formatter';
 
 import { useQuerySearchPropsWithSearchSchema } from '@/common/composables/dynamic-layout';
 import ErrorHandler from '@/common/composables/error/errorHandler';
@@ -85,7 +85,9 @@ const appContextGetters = appContextStore.getters;
 const userStore = useUserStore();
 
 const router = useRouter();
-const serviceRouter = useServiceRouter();
+const serviceRouter = useServiceRouter(router);
+
+const { referenceFieldFormatter } = useReferenceFieldFormatter();
 
 const state = reactive({
     data: undefined as any,
